Adds exit and skips argv building
I've beend having issues running compiled code under Flex. Took a
while but tracked it down to something messing with the stack. It was
the argv/argc handling. I added a bra skip and junp around the argv
code and it appears the cmds now work correctly. I'll need to revisit
the argv code to come up with something better.